% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCustomerSLA.R
\name{calcCustomerSLA}
\alias{calcCustomerSLA}
\title{Service Level Agreement determination}
\usage{
calcCustomerSLA(dataset, cfg)
}
\arguments{
\item{dataset}{The dataset with the needed modifications done by convertDate}

\item{cfg}{A json with configuration data}
}
\value{
A list with SLA, total calls and closed calls for each day of the date range specified by cfg
}
\description{
This module calculates the SLA for a specific customer
in a specific dataset
}
