% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSLA.R
\name{calcSLA}
\alias{calcSLA}
\title{Wrapper of SLA variable determination}
\usage{
calcSLA(dataset, cfg)
}
\arguments{
\item{dataset}{The dataset with the needed modifications done by convertDate}

\item{cfg}{A json with configuration data}
}
\value{
listSLA A list of lists. Each list contains the SLA for a different customer
}
\description{
Service Level Agreement is the fraction of tickets closed in time
w.r.t the total of tickets in the month until the considered time
}
