% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDate.R
\name{convertDate}
\alias{convertDate}
\title{Converts ISODate to a "numerical" format}
\usage{
convertDate(dataset, cfg)
}
\arguments{
\item{dataset}{The data frame from the source csv}

\item{cfg}{A json with configuration data}
}
\value{
dataset The same dataset but with the date/time column with the correct format.
}
\description{
The date/time of closure of tickets is in a odd format. This format disable any type of
statistical treatment.
}
