% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSLAdf.R
\name{createSLAdf}
\alias{createSLAdf}
\title{Creates dataframe simiar to a calendar}
\usage{
createSLAdf(cfg, dataset)
}
\arguments{
\item{cfg}{The json file containing all parameters.}

\item{dataset}{The dataset with the correct date object}
}
\value{
A dataframe where the rows are the days and each column is a month.
}
\description{
This function generates a data frame with the date range
given by json config file or dataset
}
