% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSQL.R
\name{createSQL}
\alias{createSQL}
\title{A SQL with the json configuration}
\usage{
createSQL(cfg)
}
\arguments{
\item{cfg}{The json configuration file read by jsonlite::fromJSON function}
}
\value{
A SQL database address (it saves the SQL too)
}
\description{
This function is used in order to create a SQL database
using the jsoin configuration input
}
