% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalls.R
\name{plotCalls}
\alias{plotCalls}
\title{Calls per day plots.}
\usage{
plotCalls(cfg, customersData)
}
\arguments{
\item{cfg}{A json with configuration data}

\item{customersData}{The dataset with the calls for all customers}
}
\value{
Do not return any output. All plots are saved in the folder defined by cfg
}
\description{
Creates the plots of calls
}
