% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSLAs.R
\name{plotSLAs}
\alias{plotSLAs}
\title{SLA plots}
\usage{
plotSLAs(cfg, customersData)
}
\arguments{
\item{cfg}{A json with configuration data}

\item{customersData}{The dataset with the SLAs for all customers}
}
\value{
Do not return any output. All plots are saved in the folder defined by cfg
}
\description{
Creates the plots of SLAs
}
