% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyRegBox.R
\name{polyRegBox}
\alias{polyRegBox}
\title{polyRegBox}
\usage{
polyRegBox(customer, variable, dataset, degree, day, verbose)
}
\arguments{
\item{customer}{Unique code associated with the customer}
}
\value{
ticketsPredictions A data frame with the number of tickets predicted
per day
}
\description{
Prediction of number of tickets using polynomial regression
}
