% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read and subset the csv}
\usage{
readData(cfg)
}
\arguments{
\item{config_json_filename}{The json file containing all parameters.}
}
\value{
A csv containing the needed data for the analysis
}
\description{
In this function the csv data source is read
}
