% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{ticketsPredictions}
\alias{ticketsPredictions}
\alias{predictions}
\title{Preview of number of calls and closed calls based on various
algorithms}
\usage{
predictions(cfg, customersData, verbose = T)
}
\arguments{
\item{cfg}{A json with configuration data}

\item{dataset}{The dataset with all customers information}
}
\value{
solution A vector with the SLA preview
}
\description{
This module receives the all calls and closed calls for a
given date range and returns the expected value of calls and closed calls
for another date range
}
