% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFolder.R
\name{setup_folder}
\alias{setup_folder}
\alias{setupFolder}
\title{Set the directory names.}
\usage{
setupFolder(cfg)
}
\arguments{
\item{config_json_filename}{A json with configuration data}
}
\description{
Set the directory names from which to load all the data for the processing
}
